third_party/blink/web_tests/platform/generic/html/tabular_data/tbody_insertrow_skips_non_tr-expected.txt

Tests that HTMLTableSectionElement.insertRow() skips non <tr> children.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS tb.__proto__ is HTMLTableSectionElement.prototype
PASS childNodes.length is 4
PASS childNodes[0].nodeValue is "TEXT"
PASS childNodes[1].tagName is "A"
PASS childNodes[2].innerHTML is "0"
PASS childNodes[3].innerHTML is "1"
PASS successfullyParsed is true

TEST COMPLETE
